refactor(frontend): split SemanticAnalyzer into semantic/ checkers
SemanticAnalyzer.cpp was a single 2743-line file: one class carrying 27
member fields plus ~58 methods that mixed symbol-table bookkeeping (the A
layer: class/function/trait/impl registration, hierarchy resolution, method
ordinals, constructor validation) with type checking (the B layer:
expression/statement/member/call checking).

The methods are now out-of-line member definitions spread across
frontend/semantic/, still members of SemanticAnalyzer so the call graph is
untouched and every `this->` / bare-field access keeps working:

- DeclCollector.cpp   registerClass/registerTrait/registerImpl/registerFunction/
                      processImpls/collectTypeAliases/resolveTypeExpr/
                      resolveDeclaredTypes/resolveHierarchy/buildMethodOrdinals/
                      validateConstructors/validateMemberInitialization/
                      findClass/classLookup            (A layer, ~443L)
- TypeResolver.cpp    substituteType/inferTypeParams/satisfiesTrait/
                      boundsSatisfied/typeFromName/typeOfValue/typeToString/
                      isNumeric/isAssignableTo/isTypeCompatible (A layer, ~212L)
- ExprChecker.cpp     checkExpr/checkExprImpl/checkAssignment/requireLet (B, ~602L)
- CallChecker.cpp     checkMethodCall/checkFuncCall/checkSuperCall/checkNew/
                      checkInstanceOf/receiverVariable/reportReceiverCaptures/
                      knownParamCount/methodSignature   (B, ~668L)
- MemberChecker.cpp   checkMemberAccess/checkMemberAccessImpl/findField/
                      findStaticField/findConstructor/findStaticMethod/
                      methodOrdinal/isDerived/staticMethodOrdinal (B, ~384L)
- StmtChecker.cpp     checkStatement/checkTopLevel/checkReturn/checkLambda/
                      checkDeclarativeBlock/checkBody/checkClassBodies/
                      checkFunctionBodies/checkClassBody/lookupName/unify/
                      isNameDefined                      (B, ~371L)

SemanticAnalyzer.cpp drops to ~87 lines (ctor + analyze orchestration).

The 27 fields stay on the class. The plan suggested hoisting them into a
separate SemanticContext and rewriting every access to `ctx.field`, but the
code reaches them by bare name ~110 times for `diagnostics` alone (vs 35
`this->` uses); a mechanical rewrite of bare identifiers would risk clobbering
same-named locals and parameters. Physical splitting of the method bodies
already hits the plan's real goal — a small main file plus a layered
directory — with zero behavioural risk, which the bytecode dump oracle
confirms (byte-identical over the 52-script corpus).

The eight namespace-level free helpers (typeFromParam/typeMatchesParam/
matchesNativeSignature/arithmeticResult/isNativeClass/basicTypes/
isReservedTypeName/isWhitelistedFormClass) that the methods call across
files now live in semantic/Helpers.cpp, declared in semantic/Helpers.h so
every checker compiles self-contained (the project's .clangd enforces
UnusedIncludes: Strict).

Behaviour unchanged: 429 tests pass, the 5 failures are the pre-existing
Bedrock-only StdlibObservableTest and the PR#35 ForInTest regression, and the
bytecode dump is byte-identical to the baseline.
